requirements.txt
aiohttp==3.8.3
aiomysql==0.1.1
aiosignal==1.3.1
anchorpy==0.15.0
anchorpy-core==0.1.2
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autoflake==2.0.0
backcall==0.2.0
base58==2.1.1
based58==0.1.1
black==22.12.0
borsh-construct==0.1.0
boto3==1.26.47
botocore==1.29.47
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
construct==2.10.68
construct-typing==0.5.5
decorator==5.1.1
exceptiongroup==1.1.0
executing==1.2.0
frozenlist==1.3.3
genpy==2022.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
ipython==8.8.0
jedi==0.18.2
jmespath==1.0.1
jsonalias==0.1.1
jsonrpcclient==4.0.2
matplotlib-inline==0.1.6
more-itertools==8.14.0
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
packaging==23.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
pyheck==0.1.5
PyMySQL==1.0.2
PyNaCl==1.5.0
pyqtgraph==0.13.1
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-xprocess==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytools==2022.1.14
rfc3986==1.5.0
s3transfer==0.6.0
shiboken6==6.4.2
six==1.16.0
sniffio==1.3.0
solana==0.29.0
solders==0.14.0
stack-data==0.6.2
sumtypes==0.1a6
tomli==2.0.1
toolz==0.11.2
traitlets==5.8.1
typer==0.7.0
types-cachetools==4.2.10
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.5
websockets==10.4
yarl==1.8.2
zstandard==0.17.0
